Fix conditional error in message_response
Actor needs to be placed in a list in order to match known_nicks, which is a list of lists.  Without that, the conditional was never matching people who joined the channel to known_nicks, which meant everyone was getting greeted.  Having known_nicks be a list of lists is kind of ugly, and would be nice to change, but for now this is a solid fix (hopefully)